refactor: move screen directory path
